pip-requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
aiofiles
python-multipart
Pillow
